Consolidated DcMotor and Servo initialization with enum constructor
Instead of defining the DcMotor names at the top in a 
enum, then defining the direction, braking, and encoder 
modes below, all this has been incorporated into the 
definition of the enum, giving a single location to 
define all motor properties.

Performed similar consolidation with the Servo enum.
Considered moving both enum definitions to the top to 
make modifications easier, but left the Servo enum in 
place for now.

Enum explanation:
http://www.discoversdk.com/blog/enums-in-java-explained