Adjusted vision position telemetry to be more readable.
Two changes: simpleVision.displayFormattedVumarkPose() 
has been made a separate function 
from updateVuMarkPose() to make the extra two lines of 
data optional.
Secondly, whether this function is included or not, the 
formatting has been adjusted so the number of telemetry 
lines displayed remains constant when vumark targets 
enter or leave the camera field of view.  
This makes the telemetry much nicer to look at.