Added CSV writer to jUnit WaypointGenerationTest.java
The .csv file is saved into the Teamcode Folder.
If you double click the .csv file, you'll be asked where 
to open it.  If you choose to associate it with an 
external program (bottom option) it will open in 
excel or open office, or whatever you have for csv's.
Then, just create a XY scatter plot from the values, 
and you have a decent visualization of the robot 
motion.