Diagnostic: Tune K_D  with dpad up/down
Changes are volatile, lost when leaving the opmode.
Every 10 increments is a factor of 10.

Added initializePID() in RobotHardware so the 
parameters displayed are up to date.

Added function for increment PIDF parameters, 
not by adding a constant, but by multiplying it.
This makes it easier to adjust parameters on a 
logarithmic scale