Corrected RGB2BGR swap, tune regions, better test images.
The default color ordering of openCV is BGR, not 
RGB. The consequence is that the files loaded for 
testing were being loaded backwards, so the 
Cb channel was backwards, and the detection 
direction was backwards.  However, the robot 
captures the camera stream in the BGR order 
assumed by openCV, causing the detections to 
be incorrect, and the saved images to have the 
red and blue channels swapped.  Corrections 
have been made for these errors.

Improved test images, captured from the robot 
camera.  Enables much better tuning.
Included some non-ideal images for more testing.

Fixed the getSkystoneRelativeLocation() function 
after the switch from BGR, the numbers now 
correctly measure blueness, instead of yellowness, 
so the direction of several tests had to be flipped.