Bugfix: Red side waypoint mirroring problem
Problem: On the Red team side, if the skystone 
is in position 1 or 2, the waypoint 
for ALIGNMENT_POSITION_A would share a 
reference with either SCAN_POSITION_A_0 or 
SCAN_POSITION_A_1.  As a result, when every 
waypoint gets mirrored in a loop, both of these 
waypoints would get mirrored twice, resulting 
in no mirroring.  Consequently, two blue team 
waypoints would get interleaved with the 
skystone portion of the red waypoints.

The solution is simply to ensure all Navigation2D 
objects assigned to the waypoint ArrayList be 
appended with .copy() so a new object be generated 
with an independent reference, to prevent 
undesired double processing.