State Refactor Template
Added final private iteration class property 
to StateBase, created when StateBase is Constructed.
Default value is 0 for StateBase(), or assigned 
with StateBase(iteration).
The new field is final, so it cannot be modified 
after the object is created. 
getIteration() allows access.

This effectively means that 'iteration' is part of 
the identity of the object, and cannot be accidentally 
changed, just as the _A or _B suffixes couldn't be 
accidentally changed.

Because iteration is final, it cannot be assigned 
directly after object creation.  You have one 
opportunity to set this value in a subclass constructor 
by calling super(iteration).

Modified the String getCurrentStates() method to 
show the iteration value, but only if it is non-zero.