Dead Wheels: Added more detailed tests, seem to be functional
Added test functions which calculate appropriate 
odometryTicks for a given Nav2D delta motion, 
so long as the delta motion is only either pure rotation 
or pure translation.
This simplifies the process of testing.
At present, this script just essentially does the 
odometry math backwards, to get ticks from delta 
position, and this serves as a simple check against typos.

-So far, the math for the center wheel offset 
compensation seems to be good.

-Most integration error is likely to occur when 
translating and rotating together.  It would be 
good to expand the ticksFromMotion code to be 
able to support other motions, although the base 
assumption is that the robot moves in arcs.