Modified Frame2D to keep (attaching to) parent frame reference
Decided that when multiple objects are in a given 
frame, they should share a common reference to that 
frame object.  If the position of that frame is modified, 
all Navigation2D objects referenced to it will also 
have their global position moved.

-Added example to Frame2D test of how to calculate 
waypoints for desired approach angles to last 
skystone, and how to verify that the wheels won't 
hit the wall.