Localizer Interface to abstract MecanumNavigation and OdometryLocalizer
Created Localizer interface to provide an abstraction 
of both the legacy MecanumNavigation and the new 
3 wheel OdometryLocalizer.  This Localizer interface 
is being used to replace MecanumNavigation 
functions in AutoDrive and PositionLocalizer, and 
most everywhere else.

Note that the preferred localizer is still constructed 
and initialized in its native way, but can then be 
passed into various other classes and used as a 
Generic Localizer transparently, regardless of its 
origin.

Note that the AutoDrive functions were overly 
coupled to MecanumNavigation, so it still needs a 
reference.  Additionally, MecanumNavigation had to 
be modified with static methods to help break this 
link.  Once we move from AutoDrive to the 
more generic PositionController, this issue will be 
gone.