PositionController - Linear and Rotation functional
-Separated PID parameters for translation and 
rotation.  Applied parameters to get reasonable 
behavior in the unit test.
-fixed bug with PositionController.toRobotFrame(),
it was correcting for the robot position, when it 
should only be concerned with the robot orientation.
-fix bug where rotation power was being modified 
while being rotated into the robot frame.
-fix bug where y and theta power sign was wrong.
-Made various internal parameters of the 
PositionController into public properties to easier 
unit testing and debugging.
-Added continuous correction for averagePeriod 
in the calculation for the per cycle ramp rate.
-Added Navigation2D.toStringPure which doesn't 
convert theta to degrees, round to tenths, or show 
units.