PID Configuration Object designed.
PositionControllerConfiguration objects just hold 
two MiniPIDCOnfiguration objects, which in turn 
hold parameters of interest for configuring the 
X, Y, and Theta PID controllers.

An additional parameter in MiniPIDConfiguration is 
positionTolerance, which is now stored in MiniPID 
and accessed within PositionController for 
arrival tolerances.

setPIDParameters(PositionControllerConfiguration) 
takes a configuration object and applies all settings 
to the 3 MiniPID controllers.  This would make it 
easier to, for example, apply a different set of 
control parameters to various waypoints.