(delta only) StdLocalizer implements Localizer, static OdoConfig for dashboard
StandardTrackingWheelLocalizer modified to use 
OdometryConfig, to access our deadwheel dcmotor 
references, and to implement the Localizer interface.

-! At present, getCurrentPosition() actually just returns 
the delta between poses, not the absolute position.
-! The current layout of directly accessing the motor 
encoder values if bad for both testing and hardware 
efficiency. The Kotlin  ThreeTrackingWheelLocalizer 
needs to be copied and modified so update() can 
accept OdometryTicks AND update() calls our 
optimized getEncoderValue() method.

-Made OdometryConfig static so it can be modified 
in real time using the dashboard.

-Added getMotor(MotorName) to RobotConfig so 
motor references could be added to the 
acme localizer.