Good Roadrunner Localization test,  needs angle unwrapped.
The test is promising, because it shows the expected 
functioning of the new Localizer. Additionally, it 
demonstrates that the deadwheel toe-in 
compensation can replicate the errors seen on 
the field, in which lateral motions would cause 
a subtle arc of about 7 degrees in 72".

To be a drop in replacement for other Localizers,
the angle will have to be unwrapped.